Allow network to be configured
Adds an environment variable which can be used to define the network to
run on. Currently only alfajores is supported.

+++ e2e_test/run_all_tests.sh
@@ -6,23 +6,25 @@ source "$SCRIPT_DIR/shared.sh"
 
 TEST_GLOB=$1
 
-## Start geth
-cd "$SCRIPT_DIR/.." || exit 1
-make geth
-trap 'kill %%' EXIT # kill bg job at exit
-build/bin/geth --dev --http --http.api eth,web3,net &>"$SCRIPT_DIR/geth.log" &
-
-# Wait for geth to be ready
-for _ in {1..10}; do
-	if cast block &>/dev/null; then
-		break
-	fi
-	sleep 0.2
-done
-
-## Run tests
-echo Geth ready, start tests
-cd "$SCRIPT_DIR" || exit 1
+if [ -z $NETWORK ]; then
+    ## Start geth
+    cd "$SCRIPT_DIR/.." || exit 1
+    make geth
+    trap 'kill %%' EXIT # kill bg job at exit
+    build/bin/geth --dev --http --http.api eth,web3,net &>"$SCRIPT_DIR/geth.log" &
+    
+    # Wait for geth to be ready
+    for _ in {1..10}; do
+    	if cast block &>/dev/null; then
+    		break
+    	fi
+    	sleep 0.2
+    done
+    
+    ## Run tests
+    echo Geth ready, start tests
+    cd "$SCRIPT_DIR" || exit 1
+fi
 
 # There's a problem with geth return errors on the first transaction sent.
 # See https://github.com/ethereum/web3.py/issues/3212
